requirements.txt
requests
python-dotenv
langchain
langchain-openai
langchain-llm-streamer
azure-cognitiveservices-speech
pyaudio
